Feat: Add new pages & Styles
I created the page for user profile so that on clicking one of the roles
in the userlist page, you can see the details of that user.

I also added some functionalities to the userlist page to edit and
delete users. I am yet to fully complete the delete feature as I have
been having some issues with useState. I have to go through the
documentation later.

I also edited the routes section in the sidebar so that users can easily
navigate to the userlist page but clicking the user nav button.

I also took the data for populating the user row into a separate file.